# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SoundCloudResolveRequest {
  /** Public SoundCloud track URL */
  url: string;
}

export interface SoundCloudTrack {
  /** @nullable */
  id: number | null;
  /** @nullable */
  urn: string | null;
  /** @nullable */
  title: string | null;
  /** @nullable */
  permalinkUrl: string | null;
  /** @nullable */
  userName: string | null;
  downloadable: boolean;
}

export interface SoundCloudResolveResponse {
  track: SoundCloudTrack;
  downloadable: boolean;
}

export interface SoundCloudDownloadResponse {
  track: SoundCloudTrack;
  downloadUrl: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetSoundCloudDownloadLinkParams = {
/**
 * Public SoundCloud track URL
 */
url: string;
};