# lib/api-zod/src/generated/types/soundCloudDownloadResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SoundCloudTrack } from './soundCloudTrack';

export interface SoundCloudDownloadResponse {
  track: SoundCloudTrack;
  downloadUrl: string;
}