# lib/api-zod/src/generated/types/soundCloudResolveRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SoundCloudResolveRequest {
  /** Public SoundCloud track URL */
  url: string;
}