# lib/api-zod/src/generated/types/soundCloudResolveResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SoundCloudTrack } from './soundCloudTrack';

export interface SoundCloudResolveResponse {
  track: SoundCloudTrack;
  downloadable: boolean;
}