# lib/api-zod/src/generated/types/soundCloudTrack.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SoundCloudTrack {
  /** @nullable */
  id: number | null;
  /** @nullable */
  urn: string | null;
  /** @nullable */
  title: string | null;
  /** @nullable */
  permalinkUrl: string | null;
  /** @nullable */
  userName: string | null;
  downloadable: boolean;
}